edit form using the modal component
