--- src/components/RoomsFormEdit/styles.css
@@ -1,45 +1,0 @@
-.roomsFormEdit {
-  padding: 1em;
-}
-
-.roomsFormEdit {
-  display: flex;
-  flex-direction: column;
-  width: 340px;
-  font-size: 15px;
-}
-
-.roomsFormEdit__title {
-  text-align: center;
-}
-
-.roomsFormEdit__properties {
-  margin: 0.5em 0 0 0;
-  font-weight: 800;
-}
-
-.roomsFormEdit__button {
-  margin-top: 1em;
-  color: white;
-  background-color: red;
-  border: 0;
-  padding: 0.5em 1em 0.5em 1em;
-  font-weight: bold;
-}
-
-.roomsFormEdit__button:hover {
-  background-color: rgb(255, 77, 77);
-}
-
-.roomsFormEdit__buttonEnv {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-@media (min-width:768px){
-  .roomsFormEdit {
-    width: 450px;
-  }
-}
-
